requirements.txt
requests
numpy
TA-Lib
python-binance
yachalk